Refactoring teams and players APIs
