+++ src/web/mod.rs
@@ -1,1 +1,2 @@
-pub mod player;
+pub mod player_rest;
+pub mod team_rest;
