+++ src/web/player_rest.rs
